pyproject.toml
[project]
name = "paper-reader"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "pymupdf4llm>=0.2.6",
    "python-dotenv>=1.2.1",
    "pyzotero>=1.7.6",
    "streamlit>=1.52.1",
]
